+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v 1.7.0 (?)
 Changes in this release:
 - Ensure that environment settings defined by Inmanta extensions are present in the generated documentation.
+- Try to load all modules provided in config options list. Do not fail is one is missing.
 
 # v 1.6.0 (2022-09-09)
 Changes in this release: